Added updated Kiuwan Reports and Analysis
The reason the old analysis was redone, was because we ran it on the Android Studio version of the code. At the time it wasn't working, but we thought we could get it to work. When we went back to eclipse, the code structure was different, so the new analysis for part 3 didn't make much sense. As such we checkout out a version of this report from before the Part 2 due date and re-ran the analysis on it